PRANAV_JAGADEESH_1BM18CS071_num1,2_threads_10b
Modify this program with Graphical User Interface
Write a program that creates a user interface to perform integer divisions. The user enters two integer in the text fields, Num1 and Num2.
The division of Num1 and Num2 is displayed in the Result field when the Divide button is clicked.
i.) If Num1 or Num2 are not integer, the program should throw a NumberFormatException.
ii.) If Num2 is Zero, the program should throw an Arithmetic Exception.
Display the exception in a message dialog box.